refactor(authflow): extract /auth request parsing into an authreq package
Move parseAuthorizationRequest, the request validation (redirect_uri,
connector_id, id_token_hint) and the request-error types (displayed vs
redirected) out of the Handler into a dedicated authflow/authreq package
(authreq.Parser). PKCEConfig moves with it — the parser owns it — and the
server and cmd reference authreq.PKCEConfig directly. The Handler holds a
*authreq.Parser and delegates parsing.
