+++ server/authflow/handler.go
@@ -22,6 +22,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/dexidp/dex/server/authflow/authreq"
 	"github.com/dexidp/dex/server/authflow/mfa"
 	"github.com/dexidp/dex/server/authflow/session"
 	"github.com/dexidp/dex/server/authflow/web"
@@ -48,7 +49,7 @@ type Config struct {
 	SkipApproval           bool
 	AlwaysShowLogin        bool
 	SupportedResponseTypes map[string]bool
-	PKCE                   PKCEConfig
+	PKCE                   authreq.PKCEConfig
 	AuthRequestsValidFor   time.Duration
 	SessionConfig          *session.Config
 	MFAProviders           map[string]mfa.Provider
@@ -61,20 +62,18 @@ type Config struct {
 type Handler struct {
 	*web.UI
 
-	connectors             *connectors.Cache
-	storage                storage.Storage
-	issuerURL              url.URL
-	templates              *templates.Templates
-	logger                 *slog.Logger
-	signer                 signer.Signer
-	issuer                 *tokens.Issuer
-	now                    func() time.Time
-	skipApproval           bool
-	alwaysShowLogin        bool
-	supportedResponseTypes map[string]bool
-	pkce                   PKCEConfig
-	authRequestsValidFor   time.Duration
+	connectors           *connectors.Cache
+	storage              storage.Storage
+	templates            *templates.Templates
+	logger               *slog.Logger
+	issuer               *tokens.Issuer
+	now                  func() time.Time
+	skipApproval         bool
+	alwaysShowLogin      bool
+	authRequestsValidFor time.Duration
 
+	// req parses and validates the /auth authorization request.
+	req *authreq.Parser
 	// sessions owns the session cookie, SSO lookup and auth-session CRUD.
 	sessions *session.Manager
 	// mfa owns the authenticator chain and the TOTP/WebAuthn endpoints.
@@ -85,22 +84,19 @@ type Handler struct {
 func NewHandler(c Config) *Handler {
 	ui := web.New(c.Templates, c.IssuerURL, c.Logger)
 	return &Handler{
-		UI:                     ui,
-		connectors:             c.Connectors,
-		storage:                c.Storage,
-		issuerURL:              c.IssuerURL,
-		templates:              c.Templates,
-		logger:                 c.Logger,
-		signer:                 c.Signer,
-		issuer:                 c.Issuer,
-		now:                    c.Now,
-		skipApproval:           c.SkipApproval,
-		alwaysShowLogin:        c.AlwaysShowLogin,
-		supportedResponseTypes: c.SupportedResponseTypes,
-		pkce:                   c.PKCE,
-		authRequestsValidFor:   c.AuthRequestsValidFor,
-		sessions:               session.New(c.Storage, c.SessionConfig, c.Now, c.Logger, c.IssuerURL),
-		mfa:                    mfa.New(ui, c.Storage, c.Templates, c.Logger, c.MFAProviders, c.DefaultMFAChain, c.Now, c.Connectors),
+		UI:                   ui,
+		connectors:           c.Connectors,
+		storage:              c.Storage,
+		templates:            c.Templates,
+		logger:               c.Logger,
+		issuer:               c.Issuer,
+		now:                  c.Now,
+		skipApproval:         c.SkipApproval,
+		alwaysShowLogin:      c.AlwaysShowLogin,
+		authRequestsValidFor: c.AuthRequestsValidFor,
+		req:                  authreq.New(c.Storage, c.Logger, c.Signer, c.IssuerURL, c.PKCE, c.SupportedResponseTypes),
+		sessions:             session.New(c.Storage, c.SessionConfig, c.Now, c.Logger, c.IssuerURL),
+		mfa:                  mfa.New(ui, c.Storage, c.Templates, c.Logger, c.MFAProviders, c.DefaultMFAChain, c.Now, c.Connectors),
 	}
 }
 
@@ -137,12 +133,4 @@ func (h *Handler) Mount(m router.Mux) {
 	m.HandleFunc("/mfa/webauthn/register/finish", h.mfa.HandleWebAuthnRegisterFinish)
 	m.HandleFunc("/mfa/webauthn/login/begin", h.mfa.HandleWebAuthnLoginBegin)
 	m.HandleFunc("/mfa/webauthn/login/finish", h.mfa.HandleWebAuthnLoginFinish)
-}
-
-// PKCEConfig holds PKCE (Proof Key for Code Exchange) settings.
-type PKCEConfig struct {
-	// If true, PKCE is required for all authorization code flows.
-	Enforce bool
-	// Supported code challenge methods. Defaults to ["S256", "plain"].
-	CodeChallengeMethodsSupported []string
 }
