+++ server/server.go
@@ -43,6 +43,7 @@ import (
 	"github.com/dexidp/dex/connector/saml"
 	"github.com/dexidp/dex/pkg/featureflags"
 	"github.com/dexidp/dex/server/authflow"
+	"github.com/dexidp/dex/server/authflow/authreq"
 	"github.com/dexidp/dex/server/authflow/mfa"
 	"github.com/dexidp/dex/server/authflow/session"
 	"github.com/dexidp/dex/server/connectors"
@@ -115,7 +116,7 @@ type Config struct {
 	PasswordConnector string
 
 	// PKCE configuration
-	PKCE authflow.PKCEConfig
+	PKCE authreq.PKCEConfig
 
 	GCFrequency time.Duration // Defaults to 5 minutes
 
@@ -211,7 +212,7 @@ type Server struct {
 
 	supportedGrantTypes []string
 
-	pkce authflow.PKCEConfig
+	pkce authreq.PKCEConfig
 
 	now func() time.Time
